Fix dev parent daemon leaks
